meetup/server/pyproject.toml
[tool.poetry]
name = "meetup"
version = "0.1.0"
description = "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
authors = ["FatJun <i.fatjuni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.1"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
aiogram = "3.0.0b7"
celery = {version = "^5.2.7"}
redis = "^4.5.5"
asyncio-throttle = "^1.0.2"
requests = "^2.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
